Add files via upload
Snake Game
The base application for the work is a Javascript implementation of the traditional Snake game. THE
encoding follows the Functional Paradigm.
The game consists of a "snake", controlled by the user through the keyboard, that moves without
interruption in four possible directions: North, South, East, West. There is an "apple" that appears
randomly on the screen, and by eating it, the snake grows in size.